% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_styles_to_mplus.R
\name{as.MplusSyntaxElements}
\alias{as.MplusSyntaxElements}
\title{Preparing Mplus code to estimate RS models}
\usage{
as.MplusSyntaxElements(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{x} with \emph{MplusSyntaxElements} class assigned
}
\description{
Assigns \emph{MplusSyntaxElements} to an object
}
