% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_from_mplus.R
\name{check_coefs_too_large}
\alias{check_coefs_too_large}
\title{Reading results from estimated Mplus models}
\usage{
check_coefs_too_large(x)
}
\arguments{
\item{x}{a data frame storing model parameters}
}
\value{
A data frame.
}
\description{
Checks if some of the columns storing the statistics
in the provided data frame contain values indicating that the statistics
value was to large to be written to the \emph{Mplus} output (making such
a column a character vector) and if so recodes such values to \code{Inf} and
transforms a column to numeric vector. Also, recodes values of \code{est},
\code{se}, \code{est_se} and \code{pval} for parameters that were not
estimated in the model (i.e. they were fixed) to \code{NaNs}.
}
