% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{generate_item_responeses_gpcm}
\alias{generate_item_responeses_gpcm}
\title{Internals: calculation of (G)PCM probabilities and drawing responses}
\usage{
generate_item_responeses_gpcm(theta, weightsMatrix, intercepts)
}
\arguments{
\item{theta}{a matrix of latent traits' values}

\item{weightsMatrix}{a matrix of slope parameters (being
multiplication of a design matrix by discriminations of factors)}

\item{intercepts}{a vector of intercept parameters}
}
\value{
a vector of responses on item
}
\description{
Computes probabilities of each response using (G)PCM and draws
responses. Be aware that for efficiency reasons it assumes  that
\strong{\code{theta} and \code{weightsMatrix} have the same number and order
of columns} (unlike \code{\link{generate_item_responses_sqn}} and
\code{\link{generate_item_responses_sml}} that are used to match theirs
arguments \code{theta}, \code{scoringMatrix} and \code{slopes} before
passing them to \code{generate_item_responeses_gpcm}).
}
