% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{generate_test_responses}
\alias{generate_test_responses}
\title{Generate simulated data}
\usage{
generate_test_responses(
  traitCov,
  items,
  n,
  traitMeans = rep(0, ncol(traitCov)),
  performAssertions = TRUE
)
}
\arguments{
\item{traitCov}{covariance matrix of traits}

\item{items}{list with test items' specification}

\item{n}{number of observations}

\item{traitMeans}{expected values of traits}

\item{performAssertions}{logical value indicating whether function should
perform assertions of the other parameters (\code{TRUE} by default, may be
changed to \code{FALSE} for a little performance gain)}
}
\value{
matrix with responses on items
\code{\link{generate_item_responses_sqn}},
\code{\link{generate_item_responses_sml}}
}
\description{
Generates simulated responses to a test given inforamations
about items and parameters of joint multivariate normal distribution of
latent traits.
}
