% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_from_mplus.R
\name{get_elapsed_time}
\alias{get_elapsed_time}
\title{Reading results from estimated Mplus models}
\usage{
get_elapsed_time(model)
}
\arguments{
\item{model}{an (already estimated) \code{mplusObject}}
}
\value{
A one-element numeric vector.
}
\description{
Extracts model estimation time in seconds.
}
