% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_from_mplus.R
\name{get_fs_reliabilities}
\alias{get_fs_reliabilities}
\title{Reading results from estimated Mplus models}
\usage{
get_fs_reliabilities(model, latentTraits)
}
\arguments{
\item{model}{an (already estimated) \code{mplusObject}}

\item{latentTraits}{a character vector with latent trait names}
}
\value{
A data frame with columns:
\itemize{
  \item{\code{latentTrait} - name of a trait,}
  \item{\code{est} - estimated reliability.}
}
}
\description{
Computes factor scores reliabilities (if available).
}
\details{
Reliability is estimated as a ratio between factor scores variance
and the sum of factor scores variance and mean squared factor scores standard
errors.
}
