% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_styles_to_mplus.R
\name{make_mplus_latent_traits_orthogonal_syntax}
\alias{make_mplus_latent_traits_orthogonal_syntax}
\title{Preparing Mplus code to estimate RS models}
\usage{
make_mplus_latent_traits_orthogonal_syntax(latentTraits)
}
\arguments{
\item{latentTraits}{a character vector of latent traits names}
}
\value{
A character vector
}
\description{
Prepares Mplus model description syntax that fixes covariances
between given latent traits to 0.
}
