% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_styles_to_mplus.R
\name{make_mplus_structural_model}
\alias{make_mplus_structural_model}
\title{Preparing Mplus code to estimate RS models}
\usage{
make_mplus_structural_model(latent, observedExogenous, observedDependent)
}
\arguments{
\item{latent}{a character vector of latent variable names}

\item{observedExogenous}{either:
\itemize{
  \item{a character vector with names of observed exogenous predictors that
        should be used to predict latent variables in the model}
  \item{a matrix with latent traits in columns and observed exogenous
        predictors in rows specifying which of the exogenous predictors should
        be used to predict which latent traits (matrix should contain only
        0 and 1 or \code{TRUE} and \code{FALSE})}
}}

\item{observedDependent}{either:
\itemize{
  \item{a character vector with names of observed dependent variables that
        should be predicted using latent variables in the model}
  \item{a matrix with latent traits in columns and observed dependent
        variables in rows specifying which of the dependent variables should
        be predicted by which latent traits (matrix should contain only
        0 and 1 or \code{TRUE} and \code{FALSE})}
}}
}
\value{
A character vector
}
\description{
Prepares Mplus model description syntax for the structural part
of the model given latent traits, observed exogenous predictors and observed
dependent variables.
}
