% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_matrices.R
\name{make_scoring_matrix_aem}
\alias{make_scoring_matrix_aem}
\title{Make scoring matrix}
\usage{
make_scoring_matrix_aem(
  responses,
  sequence = c("mae", "mea", "aem", "ame", "ema", "eam", "simultaneous"),
  nMiddle = 2L,
  nExtreme = 1L,
  nAcquiescence = floor(length(responses)/2),
  reversed = FALSE,
  iType = c("within", "between")
)
}
\arguments{
\item{responses}{vector of available responses (\emph{categories}) - can be
a character vector or positive integer describing number of responses}

\item{sequence}{sequence of decisions made by a \emph{respondent}, determined
by order of codes in a string:
\itemize{
  \item{'m' stands for choosing between middle \emph{category} and some other
        \emph{category}}
  \item{'a' stands for choosing between \emph{acquiescence} response (i.e.
        located \emph{after/below} a \emph{middle} one) and some other
        response}
  \item{'e' stands for choosing between \emph{extreme} category and some
        other \emph{category}}
}}

\item{nMiddle}{(maximum) number of \emph{middle} \emph{categories}}

\item{nExtreme}{(half of the) number of \emph{extreme} \emph{categories}}

\item{nAcquiescence}{number of \emph{acquiescence} \emph{categories}}

\item{reversed}{logical value - is item a reversed one? (see details)}

\item{iType}{determines a way in which scoring pattern for additional
\emph{intensity} trait will be generated (see details)}
}
\value{
matrix of integers
}
\description{
Make scoring matrix
}
\details{
\strong{\code{sequence} other than "simultaneous"}

For number of responses between 5 and 6 function generates scoring
matrix in a way mimicking Böckenholt's approach (2017) to describe
response to the item as a sequence of binary decissions involving choosing
of the middle, extreme and acquiescence categories - this decisions may be
made in diferent order, what is controlled by parameter \code{sequence}.

With less than 5 possible responses functions apply the same logic, but not
all of the three aformentioned styles can be invloved because lack of
variablity in possible responses.

With more than 6 possible responses there must be odditional trait added to
scoringMatrix to describe process of choice between all the possible
responses. In such a case function adds 4th column to a scoring matrix that
is named \emph{i} (standing for intensity) and is filled up with scores for
such rows of a scoring matrix that occurs more than once. Scores in this
column are sequences of positive integers either increasing
(\code{reversed=FALSE}) or decreasing (\code{reversed=TRUE}) that are
generted either as a one sequence for all the rows to be filled up
(\code{iType="between"}) or as independent sequences for each unique
combination of values in the precious columns (\code{iType="within"}).)

\strong{\code{sequence} is "simultaneous"}

In this case a GPCM scoring matrix is generated mimicking approach of
Plieninger (2016), i.e. assuming that response process is
a \emph{simultaneous} and four factors: intensity of the trait that
is \strong{not} a response style (column \emph{i}), tendency to choose middle
\emph{categories} (column \emph{m}) tendency to choose extreme
\emph{categories} (column \emph{e}) and tendency to choose cquiescence
\emph{categories} (column \emph{a}) contribute alltogether to propensity
of choosing each response.
}
