% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_matrices.R
\name{make_scoring_matrix_rt}
\alias{make_scoring_matrix_rt}
\title{Make scoring matrix}
\usage{
make_scoring_matrix_rt(responses)
}
\arguments{
\item{responses}{vector of available responses (\emph{categories}) - can be
a character vector or positive integer describing number of responses}
}
\value{
matrix of integers
}
\description{
Makes response matrix using \emph{random thresholds} approach.
}
\details{
Be aware that while using this kind of response matrix latent
traits must be set orthogonal to assure model identifiability.
}
\examples{
make_scoring_matrix_rt(5)
}
