% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_matrices.R
\name{make_scoring_matrix_stz}
\alias{make_scoring_matrix_stz}
\title{Make scoring matrix}
\usage{
make_scoring_matrix_stz(responses)
}
\arguments{
\item{responses}{vector of available responses (\emph{categories}) - can be
a character vector or positive integer describing number of responses}
}
\value{
matrix of integers
}
\description{
Makes response matrix using \emph{sum to zero} approach.
}
\examples{
make_scoring_matrix_stz(5)
}
