% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_styles_to_mplus.R
\name{print.MplusSyntaxElements}
\alias{print.MplusSyntaxElements}
\title{Preparing Mplus code to estimate RS models}
\usage{
\method{print}{MplusSyntaxElements}(x, ..., n = 10L)
}
\arguments{
\item{x}{an object of class \emph{MplusSyntaxElements}}

\item{...}{optional arguments to \code{print.data.frame} methods (used only
if \code{x} has an element that is a data frame)}

\item{n}{an integer passed to \code{\link{head}} (used only if \code{x}
has an element that is a data frame)}
}
\value{
invisibly \code{x}
}
\description{
Print method for objects containing Mplus syntaxes
}
