% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_from_mplus.R
\name{read_mplus_object_results}
\alias{read_mplus_object_results}
\title{Reading results from estimated Mplus models}
\usage{
read_mplus_object_results(
  model,
  wide = vector(mode = "character", length = 0L),
  deleteDataFile = TRUE
)
}
\arguments{
\item{model}{an (already estimated) \code{mplusObject}}

\item{wide}{optionally a character vector indicating that resulting data
frames should be transformed to \emph{wide} format with only one row and
all the statistics in separate columns; argument gives names of the
statistics - to be chosen from the set: \code{"est"}, \code{"se"},
\code{"est_se"}, \code{"pval"} - that should be included in the resulting
data frames}

\item{deleteDataFile}{a logical value indicating whether to remove from disk
the text file storing the data from which the model was estimated}
}
\value{
A list with elements:
\itemize{
  \item{\code{fitIndices} - one-row data frame containing model fit indices
        and possibly warnings an errors,}
  \item{\code{pathCoefs} -  a list with element \code{unstandardized} and
        possibly other elements storing standardized results; each of the
        elements is a data frame storing slopes and intercepts from the
        structural part of the model and also possibly latent trait means
        (if freely estimated in the model),}
  \item{\code{variances} - a list with element \code{unstandardized} and
        possibly other elements storing standardized results; each of the
        elements is a data frame storing variable variances or residual
        variances,}
  \item{\code{covariances} - a list with element \code{unstandardized} and
        possibly other elements storing standardized results; each of the
        elements is a data frame storing variable covariances,}
  \item{\code{slopes} - a list with element \code{unstandardized} and
        possibly other elements storing standardized results; each of the
        elements is a data frame storing slope/discrimination parameters
        linking latent traits with their items (observed indicators),}
  \item{\code{thresholds} - a list with element \code{unstandardized} and
        possibly other elements storing standardized results; each of the
        elements is a data frame storing threshold/step/intercept item
        (observed indicators of latent traits) parameters,}
  \item{\code{rSquared} - R-squared statistics for observed dependent
        variables (if reported in the \emph{Mplus} output),}
  \item{\code{reliabilities} - factor scores reliabilities, available only
        if factor scores and their standard errors were estimated.}
}
}
\description{
Transforms results included in estimated \code{mplusObject}
into list of (lists of) data frames storing different sets of parameters.
}
