% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds_and_intercepts.R
\name{thresholds_and_intercepts}
\alias{thresholds_and_intercepts}
\alias{thresholds2intercepts}
\alias{thresholds2intercepts.data.frame}
\alias{thresholds2intercepts.matrix}
\alias{thresholds2intercepts.default}
\alias{intercepts2thresholds}
\alias{intercepts2thresholds.data.frame}
\alias{intercepts2thresholds.matrix}
\alias{intercepts2thresholds.default}
\title{Conversion between thresholds and intercepts parametrisations}
\usage{
thresholds2intercepts(thresholds)

\method{thresholds2intercepts}{data.frame}(thresholds)

\method{thresholds2intercepts}{matrix}(thresholds)

\method{thresholds2intercepts}{default}(thresholds)

intercepts2thresholds(intercepts)

\method{intercepts2thresholds}{data.frame}(intercepts)

\method{intercepts2thresholds}{matrix}(intercepts)

\method{intercepts2thresholds}{default}(intercepts)
}
\arguments{
\item{thresholds}{vector or matrix (thresholds in cols, items in rows) of
item thresholds}

\item{intercepts}{vector or matrix (intercepts in cols, items in rows) of
item intercepts}
}
\value{
vector or matrix of thresholds or intercepts6
}
\description{
Utility functions allowing to covert \emph{thresholds} (i.e.
parameterisation convienient to think of item steps difficulty) to
\emph{intercepts} (i.e. parameterisation used internally by packages
\emph{rstyles} and \emph{mirt}) or back forth.
}
\examples{
# 5 items with (general) difficulty evenly spanned between -2 and 2
# and for each item thresholds eveny spanned betwenn -1.5 and 1.5 relatively
# to item difficulty
thresholds <- matrix(rep(seq(-2, 2, by = 1), 4) +
                     rep(seq(-1.5, 1.5, by = 0.75), each = 4),
                     ncol = 4)
(intercepts <- thresholds2intercepts(thresholds))
intercepts2thresholds(intercepts)
# works also for vectors
thresholds2intercepts(thresholds[1, ])
intercepts2thresholds(intercepts[1, ])
}
\seealso{
\code{\link{generate_intercepts}}
}
