% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{generate_test_responses}
\alias{generate_test_responses}
\title{Generate simulated data}
\usage{
generate_test_responses(
  theta,
  items,
  performAssertions = TRUE,
  tryConvertToNumeric = TRUE
)
}
\arguments{
\item{theta}{a matrix (or data frame) of already generated latent traits'
values}

\item{items}{a list with test items' specification}

\item{performAssertions}{a logical value indicating whether function should
perform assertions of the other arguments (\code{TRUE} by default, may be
changed to \code{FALSE} for a little performance gain)}

\item{tryConvertToNumeric}{a logical value indicating whether function should
try to convert generated responses to numeric values (technically responses
are get from rownames of the items' scoring matrices that are character
vectors)}
}
\value{
matrix with responses on items
\code{\link{generate_item_responses_sqn}},
\code{\link{generate_item_responses_sml}}
}
\description{
Generates simulated responses to a test given informations
about items values of latent traits.
}
