% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_responses.R
\name{generate_test_responses}
\alias{generate_test_responses}
\title{Generate simulated data}
\usage{
generate_test_responses(theta, items, performAssertions = TRUE)
}
\arguments{
\item{theta}{matrix (or data frame) of already generated latent traits'
values}

\item{items}{list with test items' specification}

\item{performAssertions}{logical value indicating whether function should
perform assertions of the other arguments (\code{TRUE} by default, may be
changed to \code{FALSE} for a little performance gain)}
}
\value{
matrix with responses on items
\code{\link{generate_item_responses_sqn}},
\code{\link{generate_item_responses_sml}}
}
\description{
Generates simulated responses to a test given informations
about items values of latent traits.
}
